bind picked kana to game
