+++ src/lib/quiz.ts
@@ -1,6 +1,4 @@
-import type { GameSettings } from "@/stores/settings"
 import { shuffleArray } from "./random"
-import { getDictionary } from "./dictionary"
 
 export interface QuizItem {
   index: number;
@@ -16,10 +14,6 @@ export function quizItem(kana: string, i: number): QuizItem {
   }
 }
 
-export function getQuiz(opts: GameSettings): QuizItem[] {
-  const dictionary = getDictionary(opts);
-
-  const items = shuffleArray(dictionary).map(quizItem);
-
-  return items;
+export function getQuiz(dictionary: string[]): QuizItem[] {
+  return shuffleArray(dictionary).map(quizItem);
 }
